+++ app/src/main/java/org/retriever/dailypet/ui/signup/profile/TermOfServiceFragment.kt
@@ -1,11 +1,9 @@
-package org.retriever.dailypet.ui.signup.firstcourse
+package org.retriever.dailypet.ui.signup.profile
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.CheckBox
-import android.widget.Toast
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.asLiveData
