+++ app/src/main/res/layout/fragment_create_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.signup.firstcourse.CreateProfileFragment">
+    tools:context=".ui.signup.profile.CreateProfileFragment">
 
     <com.google.android.material.progressindicator.CircularProgressIndicator
         android:id="@+id/progress_circular"
