+++ app/src/main/res/layout/fragment_term_of_service.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.signup.firstcourse.TermOfServiceFragment">
+    tools:context=".ui.signup.profile.TermOfServiceFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/left_guideline"
